@@ -5,14 +5,14 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9.0,<4.0.0"
+requires-python = ">=3.10.0,<4.0.0"
 dependencies = [
     "qdrant-client>=1.10.1,<2.0.0",
     "pydantic>=2.7.4,<3.0.0",
-    "langchain-core!=0.3.0,!=0.3.1,!=0.3.10,!=0.3.11,!=0.3.12,!=0.3.13,!=0.3.14,!=0.3.2,!=0.3.3,!=0.3.4,!=0.3.5,!=0.3.6,!=0.3.7,!=0.3.8,!=0.3.9,<0.4.0,>=0.2.43",
+    "langchain-core>=1.0.0a6,<2.0.0",
 ]
 name = "langchain-qdrant"
-version = "0.2.1"
+version = "1.0.0a1"
 description = "An integration package connecting Qdrant and LangChain"
 readme = "README.md"
 
@@ -49,9 +49,6 @@ typing = ["mypy>=1.10.0,<2.0.0", "simsimd>=6.0.0,<7.0.0", "langchain-core"]
 langchain-core = { path = "../../core", editable = true }
 langchain-tests = { path = "../../standard-tests", editable = true }
 
-[tool.ruff]
-target-version = "py39"
-
 [tool.ruff.format]
 docstring-code-format = true
 
@@ -68,6 +65,7 @@ ignore = [
     "UP045",   # pyupgrade: non-pep604-annotation-optional
     "PLR0913", # Function has too many arguments
     "C901",    # Complex functions
+    "TC003",
 
     # TODO"
     "ANN401",
