@@ -18,7 +18,7 @@ def qdrant_running_locally() -> bool:
 def assert_documents_equals(actual: list[Document], expected: list[Document]) -> None:  # type: ignore[no-untyped-def]
     assert len(actual) == len(expected)
 
-    for actual_doc, expected_doc in zip(actual, expected):
+    for actual_doc, expected_doc in zip(actual, expected, strict=False):
         assert actual_doc.page_content == expected_doc.page_content
 
         assert "_id" in actual_doc.metadata
